FullTube Bug Fix

+++ src/controllers/videoControllers.js
@@ -17,19 +17,31 @@ export const watchVideo = async (req, res) => {
 
 export const getEditVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await videoModel.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "404 Error" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEditVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await videoModel.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "404 Error" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await videoModel.findByIdAndUpdate(id, {
     title,
@@ -71,7 +83,17 @@ export const postUploadVideo = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
+  const video = await videoModel.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "404 Error" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await videoModel.findByIdAndDelete(id);
   return res.redirect("/");
 };
